requirements.txt
websockets==6.0
bitshares==0.2.1
websocket_client==0.56.0
